requirements.txt
torch
torchvision
matplotlib
seaborn
scikit-learn
Pillow
gradio==4.44.1
gradio_client==1.3.0
huggingface-hub==0.24.7
pydantic==2.8.2
timm
pandas
